Offers limit parameter and offers double Quality cast https://github.com/Transia-RnD/xrpl.c/issues/11

+++ Xrpl.C/Xrpl/Client/Model/Account/AccountOffers.cs
@@ -43,7 +43,7 @@ namespace Xrpl.Client.Model.Account
         public Currency TakerPays { get; set; }
 
         [JsonProperty("quality")]
-        public uint Quality { get; set; }
+        public double Quality { get; set; }
 
         [JsonProperty("expiration")]
         [JsonConverter(typeof(RippleDateTimeConverter))]
